Miasmata: Use loop to improve stereo cursor accuracy
The idea here is to start at the center of the screen (screen depth) and
iterate towards center + separation (infinity) testing the depth at each
point to try to more accurately find the point that the cursor is
actually over.

This can't use the depth buffer from both eyes, so this will still be
inaccurate around the egdes of objects. There also seems to be some
situations where this is not behaving quite as I expect (crosshair
slightly in front or behind of where I would expect for cases I thought
would be accurate), so I may still have an issue here.