Miasmata: Fix stereo cursor on water at edge of map
On the water right at the edge of the map the depth buffer is blank (0),
which moves the stereo crosshair to screen depth. 0 is never valid
(would be a divide by zero in the perspective divide if it was), so
ignore it.

My guess is this is done in the game so the water at the edge of the map
is rendered fully opaque.