BoUT2: Use different vertex shader to match pause menu book
This seems to be a bit more reliable and works in some scenes where the
other one was not (e.g. Village in chapter 5). I think the other one was
adding additional shading and not the base texture.

For now I've removed the other one, but if this also turns out to be
unreliable I could try using both.