DreadOut: Detect if cameraphone is active for shadow correction in Act 2
This game has been tricky with it's dual FOV thanks to the camera phone.
Up until now I'd been able to find ways to fix the shadows for both
reliably, but Act 2 has thrown that out the door.

This is really where we need a wrapper that can filter based on render
target, and some more smarts in matrix copies (e.g. filtering out
the identity matrix and filtering on matching render targets) would help
a lot.

I was eventually able to find a way to match when the camera phone is
active (Helix mod was also causing trouble here), which allows me to
switch between two correction factors, and will have to do for now.