DreadOut: Test light matrices for identity to determine when they are valid
If the light matrices are invalid they will be the identity matrix. I
test MVP[3,3] == 1 to determine if it is the identity matrix. I don't
test the MV.I matrix as it will have floating point errors making the
test unreliable.

If the matrices are invalid we use predefined correction values based on
whether the camera is being drawn or not. The problem with this is the
shadows outside the camera will be broken, but there's not much else I
can do with the features currently available in Helix mod.