DreadOut: Fix shadows in cut scene near beginning of Act 2
Yet another case where the existing rules did not work :(

In this case it turned out that I could use the FOV from the non-light
shaders as the camera is inactive in the cutscene and these give an
accurate FOV.

So, the rules for directional lighting are now:

If the matrices copied from PrePassLighting are not the identity matrix,
use them.

Otherwise, if the camera is active, use a hard coded constant.

Otherwise, if the camera is inactive, use the matrices copied from other
shaders (no test for identity matrix yet).