DreadOut: Note unfixable shadows in some cutscenes in Act 2
Specifically hit this on the second cutscene when barricading a door.
Matrices are available with the correct FOV (and other shadows in this
cutscene are fixed), but there seems to be no way to get them in the
current Helix Mod.

This would be easy, if only Helix Mod ignored Identity matrices.