Oddworld: Add convergence preset for pause menu
This shader is used for the transparent region of the pause menu. With
any luck this one won't be used in game, but I'm not holding my breath.

I don't think there is anything else in the menu that I can match
against, so if this doesn't work out the menus will just have to be
without a preset.