Oddworld: Fix ripples around Abe with red power circles
The shockwave ripples are a post-processing effect without access to the
necessary depth information to correctly align the ripples. Copy part of
the MVP matrix from the shader for the red circles around Abe, and use
it to calculate the depth by running a coordinate of the origin through
it.