Oddworld: Fix sun shafts at top of Scrab temple
Looks like the same shader as Stranded Deep, but here the fix was only
necessary in the Vertex shader.

Keep an eye out for more variants in case this is a repeat of Stranded
Deep - so far the end of the scrab temple is the only place I've even
seen these.