Stranded Deep: Automatically remove crosshair when faded out
The crosshair is detected based on both the texture coordinate, and
vertex position to minimise the chance that this will get applied to
something else.

The colour of the crosshair is checked to determine if it is faded out
and (not over an object of interest) and it is removed in that case.