Far Cry 4: Re-enable bloom and push dirty lens effect back
Bloom looks ok in 3D this game, but the dirty lens is at screen depth.

Re-enable the bloom and adjust the coordinates of the dirty lens to push
it back to depth, stretching it to avoid clipping at the edge of the
screen (alternatively the dirty lens could be disabled leaving bloom,
but I've run out of ini params to make it user customisable).