Far Cry 4: Adjust weapon preview depth in menus to counter UI depth
These are rendered to a intermediate buffer which is then rendered to
part of the UI, and as such they can appear beyond infinity if the UI is
pushed back. Adjust their position to Cancel out the UI depth adjustment
and render them at the original depth.

The background of this render target does not get cleared in the left
eye for me :-/