Far Cry 4: Fix reflected trees (more or less)
Some of the tree reflections were floating randomly where they did not
belong - looks like this shader was broken and misplaced them.

Fixed this from scratch since there was a few things wrong, and the
shader had a lot of decompile issues. I still haven't managed to get it
quite right - it selects the wrong imposter from the array, so
reflections don't line up as nice as they should, and the imposter is
missing from some angles. I've got it pretty close to the original
assembly and I still can't see what's wrong. Probably close enough for
now - at least the reflection is in the right location now, and there's
only a few places in the game where it's even noticeable that it's
picked the wrong imposter...

To compare with the original shader it's best to turn off 3D, find
somewhere with distinctive tree reflections (The lake in Shangri-La is
quite good for this) and use the show_original key binding to compare.

Could be a rounding error - I noticed the PI constant was truncated,
maybe something else was too? Not sure, just guessing.