Far Cry 4: Fix fog (e.g. in Shangri-La temple)
This is another effect that done by the same shaders that are also
responsible for specular reflections, environment map and general
lighting (but not shadows).

Pretty simple - I already use both corrected and uncorrected world
coordinated in the shader, the fog was using what I had assumed was a
temporary variable further down and needed the adjusted version, so it
was just a matter of switching them around.