Far Cry 4: Fix flickering shadow due to floating point error
Near the start of Act 3 one of the light casts shadows on two nearby
tables that had bad flickering. It's likely that that particular light
had a very low (or zero) shadow bias set and that combined with the
floating point error we were introducing through the shadow fix was
enough to cause severe flickering artefacts.

Restructure the maths to keep the floating point values lower, which
better utilises their available precision, reducing error and
eliminating the flickering.