matrix.py: Work out the maths to do (MV.I * MVP)[0,0].I in assembly
Since Helix mod only has two slots to copy matrices, it would be nice if
we could avoid using one of them to invert the MV matrix when we only
need it in the shader it came from (as I do in the Unity lighting fix
template).

Calculating a general matrix inverse in assembly would be pretty nasty,
but it turns out that we can make enough assumptions and simplifications
that this is actually fairly straight forward, requiring only 15
instructions.