Unity template: Invert the MV matrix inside the lighting shader
This implements the maths worked out in the previous commit to invert
the MV matrix inside the lighting shader to avoid needing helix mod to
do it for us. Helix mod is still used to invert matrices passed between
shaders, but we no longer need it to invert the matrix only used back in
the original shader, freeing up a matrix copy slot for other purposes.

Tested on Oddworld