Dreamfall Chapters: Update PrePassLighting shaders to resemble latest template
I've made a number of improvements to the lighting fix since originally
fixing Chapter 1, and I've found one area in Chapter 2 that complicates
things, so start by adding the improvements from the template to the
variant of PrePassLighting that this game uses.

- Update VS92E51A17 to check if drawing a full screen quad or light in
  the scene and use different logic to calculate the FOV, like the
  template.
- Pass the calculated FOV to the pixel shaders
- Symlink pixel shaders from template that match
- Update other PrePassLighting pixel shaders to remove uncorrection on
  full screen variants, and to use passed in FOV rather than calculating
  it from scratch
- Add an additional FOV source (just one for now)

PrePassCollectShadows is unchanged beyond the FOV source now coming from
a different shader, as the Dreamfall version differs too much from the
original.