Dreamfall Chapters: Add more FOV sources
With only the particle blend shader, the directional shadows would lift
off the ground, suggesting that it was only used after the shadows had
been drawn in the frame, so the directional lighting shader got outdated
FOV information. Add a few more FOV sources which seems to have done the
trick.