Eliminate one mov instruction from matrix inverse assembly
There is absolutely no reason to do this, other than being able to say
that I've done the full calculation including a matrix inverse in 14
lines of assembly the next time I'm at a game dev meetup.

I could make that 13 lines if I replaced two adds with a dp3, but I'm
not sure the cost of that is worth saving one line.