matrix.py: Add projection matrices that include the nv adjustment
I'd heard that the nvidia stereo correction was equivalent to an
off-center projection, but when I ran the maths that just didn't work
out.

Instead, I came up with a different modified projection matrix that is
equivalent to using a regular projection matrix then adjusting the
output coordinate with the nvidia stereo correction formula.

This might be useful to integrate nvidia 3D Vision support into game
engines as a way to switch out their projection matrix (since at least
in UE4 this is already pluggable for VR support) and maintain the same
projection we are used to with 3D Vision automatic.

On the other hand, this is not quite the same as an off-center
projection matrix, so some optimised matrix maths that assumes that
Projection[3,0] == 0 may not work with it.