Viscera: Fix & adjust UI
The UI was already being stereo corrected, but so much so that it just
disappeared with the separation much above 1%.

The UI vertex shader is also used for some kind of full screen effect,
to tie the adjustment to the pixel shader.

Uncorrecting the UI vertex was no good, as doing so broke other parts of
the UI. Instead, perform the perspective divide in the shader, removing
the depth information that the driver was using, then insert our own
depth adjustment.

Also adjust the mouse cursor depth, which is just a standard UI depth
adjustment.