game list: Reduce maxResults to 500
Words are not enough to describe how pissed I am at Google for breaking
this. Hey @Google, can't you send a message to the owner of any API keys
you are going to break BEFORE you do so!?!

500 will cover us for a little while, but not for long. I'll have to try
using pageToken to break the request up into multiple pages, but the
whole reason I had maxResults so high in the first place was because
pageToken never worked.