extract_unity_shaders: Match 3Dmigoto's (incorrect) "FNV" hash function
The hashes this was producing did not match 3Dmigoto's. On closer
inspection I realised that 3Dmigoto was not correctly seeding the hash
function.

Fixing this in 3Dmigoto might be painful as this would break all
existing fixes unless we include some backwards compatibility flag, so
for now just match 3Dmigoto's hash function instead.