Life is Strange: Disable fog broken by SkipSetScissorRect and/or update
This is strange... The chapter 2 update broke fog in a away that looked
like scissor clipping, but only when SkipSetScissorRect is enabled :-/

It's almost like the game has managed to set the scissor rectangle
anyway, but SkipSetScissorRect prevents it from clearing it to draw fog?

Oh Helix mod, why couldn't you be open source?

I need SkipSetScissorRect to fix decals, so kill the fog instead. P can
be used to toggle the fog back on.