Unity 5 DX11: Fix directional shadows
This seems easier than the equivalent Unity 4 DX9 shaders, since it
already has access to the necessary inverse projection matrix to fix the
shadows, and otherwise the relevant transformations are pretty much
unchanged.

Use unity_asset_extractor and extract_unity_headers to get the headers
and determine which constant buffer is which, then grab the original
structure from the source code for the shaders.

Pass the relevant part of the inverse projection matrix through to the
pixel shaders where it can be used to correct the shadows.