Unity 5 DX11: Fix point lights
Again this follows much the same pattern as Unity 4.

The inverse model view matrix supposedly passed in from Unity seems to
be 0, so we inverse enough of the model-view matrix to calculate what we
need from the inverse projection matrix instead (like we did for Unity
4). This matrix inverse is done in the shader so there is no need to add
support to 3Dmigoto for it.

Note that this now seems to use Hidden/Internal-DeferredShading in
addition to (or instead of?) the old Hidden/Internal-PrePassLighting
shader, though they both share the same vertex shader.