Unity 5 DX11: Partial fix for physical lighting (NEEDS MATRIX COPY!!!)
The pattern for physical based lighting (i.e. specular reflections, etc)
is the same as regular lighting, but it is applied over the full screen,
so the MV and MVP matrices in the lighting shader are not valid and
cannot be used to determine the FOV.

This means we will need to implement support for copying matrices and/or
constant buffers in 3Dmigoto to make this fix FOV correct.

For now just set inverse_projection._m00 to 1 as an approximation.