Unity 5 DX11 Viking Village: Fix sun flare from confused driver heuristics
Like Unity 4 this shader matched several things, including GUI so this is
likely no the final fix (The Viking Village sample only used this for
the sun flare).

A new twist was that the driver inconsistently applied the stereo
correction to this effect, so that sometimes it is at screen depth and
sometimes it is in the distance near infinity (not quite skybox depth,
but not far off it).

In order to solve this issue, normalise the position such that it will
be at convergence depth, and therefore not adjusted by the driver
regardless of it's heuristics, freeing us to apply our own stereo
correction with consistent results.