Unity 5 DX11 Viking Village: Fix DoF blur in level fly-over
Press C in the sample to start the fly-over, which enables motion blur
and DoF blur. The highlights in the DoF blur were being drawn in both
eyes.

It turned out that one shader was appending the location of highlights
to a structured buffer, which was being read from a separate shader and
drawn on the screen. Since structured buffers cannot be stereoised like
textures, the final shader received points for both eyes in both eyes,
while it only should process points for the current eye.

In order to fix this I encoded the eye into each point, by negating
pos.z and subtracting 1 for the right eye. In the second shader I was
then able to filter out points for the wrong eye.

Both of these shaders failed to decompile, but since I have access to
their source code I decided to experiment with using the Unity Cg source
directly. In order to do this:

- I copied the vert() or frag() functions from the Cg source, renaming
  them to main()
- I added all necessary structure definitions and called functions from
  the Cg source
- I used the extracted headers to determine which constant buffer was
  which, and the pack offsets of each global variable (divide by 16!)
- I used the extracted headers to determine which Tex2D and sampler
  registers were associated with which texture (the Tex2D register is
  the number after SetTexture, while the sampler register is the number at
  the end of the line after 2D, and these may not always match)
- I translated the tex2D sampling method from Cg into the equivalent
  DX11 method.