unity_asset_extractor: Update to handle version 15 asset files in The Forest
TOC entries now take 28 bytes. Not sure what the new 4 bytes are for -
mostly 0, but final entry may be different. Looks like there may be a
new table after the TOC, which it may be related to.

Also identified one previously unknown byte as being part of the unknown
table length. u6+u7 probably also part of it, but I'll wait until I find
a game that triggers an assert to confirm.