The Forest: Add Unity 5 DX9 Hidden/Internal-PrePassCollectShadows shaders
I've previously determined that these are responsible for directional
lighting in Unity 5, as they were in Unity 4. Add the original shaders
for diffs.

Since it is likely that The Forest is still using custom lighting
shaders, don't add these to the template yet until I can determine which
are from vanilla Unity 5.