shadertool: Add --fix-unity-lighting-ps pattern
This applies part of the Unity lighting fix - specifically, the
adjustment performed in the pixel shader. It scans the pixel shader and
attempts to identify the registers with the depth and x coordinate, then
inserts an adjustment.

The vertex shader cannot be automatically fixed, but there are only a
few variants and they sometimes need game specific tweaks, so using a
template and tweaking as necessary is easier for that case.

The vertex shader must store part of the inverse projection matrix in
texcoord 5 for use in the pixel shader. It is left up to the vertex
shader to get this in any way it can, or hard code it if all else fails.