shadertool: Raise exception if adding an explicit stereo sampler that is already used
If the sampler has not been specified it will try to use the default, or
find a free one and report the lines that need to be added to the ini.

On the other hand, if a specific sampler had been specified that was
already used in the shader, it would add a second declaration for that
sampler, which is illegal. Raise an exception instead.