The Forest: Fix Unity 5 deferred lighting shaders
This uses the inverse projection matrix copied from the directional
lighting shader. If valid, this may be more reliable than the old method
used to determine it in Unity 4.

Will probably still need to bring back the Unity 4 techniques for
situations where directional lighting is not in use.

NOTE: This has not been tested