The Forest: Add possible alternate FOV sources
One key difference from Unity 4 is that the MV matrix is only three
columns (4th is implied 0001)! This means that using InverseMatrix in
Helix mod will not do the right thing as it expects a 4x4 matrix, but
will have some unrelated vector in the 4th column.

This should be fairly easy to address with a little restructuring in the
vertex shaders to perform the inversion there instead.