The Forest: Fix inverse MV calculation for 3x4 matrix (+ other bug)
Since Unity 5 now uses 3x4 matrices for the MV matrix, we can't use
Helix mod to invert it like we could in Unity 4. Instead we use the code
I've written previously to inverse it in the shader, since that code
already is built on the assumption that the 4th column is 0001.

Also fix a couple of constants I missed updating when moving this to
Unity 5.