extract_unity_shaders: Optimise curly_scope() path
The World of Diving Unity 5 beta has a 57MB (!) shader, which is
bringing out the bottlenecks in my parsing code.

Rework curly_scope to remove the pop() call and minimise memory
manipulations - this improves the speed of that function by at least
several orders of magnitude, taking seconds instead of tens of minutes.