extract_unity_shaders: Add --type argument to filter processing shader types
This option is handy to avoid calculating the CRC32 of DX9 hashes (which
is slow due to calling out to an external tool and assembling the
shaders) when we only need the DX11 shaders (or vice versa).

This can reduce the time taken to process the 57MB (!)
StandardSpecular.shader in the Unity 5 World of Diving beta to under a
minute, down from tens of minutes.