shader database: Use ctime rather than atime to check for updates
atime can be updated by simply reading a file (or not, if the filesystem
was mounted with noatime), so it's not ideal to use to check for
updates.

mtime is also no good since we want it to match the last modified time
of the downloaded fixes.

That leaves either ctime, or writing out a status file.

ctime won't generally be updated accidentally (though it will be updated
by a chown, chmod, etc), and we can update it by touching the file
permissions after checking for an update, so use that.