Dead or Alive 5: "Fix" clipping on flashing health bar during crit
The region of the health bar that flashes is set via the scissor
rectangle (no wonder I couldn't find the shader).

Disable the scissor rectangle to fix the clipping. Unfortunately this is
not ideal as the entire health bar now flashes, but that still looks
better than the alternative.

If only Helix mod was open source I could have it pass the scissor
rectangle coords into the shader and adjust it properly...

Also fixes clipping on some of the tutorial messages :)