Tinker: Brand new shadow fix using frustrum and far clipping plane!
Since this game does not reliably have any FOV sources that we would
typically use in a Unity game, I've come up with an alternative method.

The fog shaders in this game have the camera frustrum (and so far seem
to be consistently used...). We can take the top left and top right
point on the frustrum to get the horizontal distance, then divide that
by the far clipping plane to determine the ratio of the horizontal FOV,
equivelent to tan(fov_horizontal / 2)!

The far clipping plane is fairly easy to get in Unity games as it is in
_ProjectionParams.z, and it's reciprocal is in _ProjectionParams.w - so
just copy that from other shaders.