mkrelease / Euro Truck Simulator 2: Add support for complex zip file structures
This allows me to mirror the directory structure of some games in the
zip file. If a 'zip' directory exists in a fix folder, it is used
instead of the main folder.

Files can be symlinked from the parent folder, which allows
ShaderOverride and DX9Settings.ini to be duplicated easily for
32bit/64bit with no risk of one copy becoming stale and such that
shadertool.py --git can still be used.

If the README.md file exists, the renamed version will always be added
to the root of the zip file.