The Long Dark: Add toggle for interior volumetric light shafts
These are the same light shafts I looked at in Dreamfall Chapters, where
I eventually concluded that they were unfixable (without modifications
to the C# code to pass the depth of the light source into the shader),
so kill them by default and add a toggle key.

Notably, the non-full screen variants appear to position the shafts
correctly (which IIRC did not happen in Dreamfall), so it might be worth
investigating these again. The depth buffer is still being sampled at
the wrong coordinates though, so for now they are still broken.