The Long Dark: Match rest of UI to crosshair depth
Like Stranded Deep, the motivation here is to have item names at the
right depth. For now this moves the whole UI to the crosshair depth - it
might be possible to filter such that only the text of interest moves,
but in Stranded Deep that proved rather difficult (if not impossible).