Montague's Mount: Holy shltballz it works! (MakeWTexture / DepthWRT / DepthWReg)
My theory as to why copying _CameraDepthTexture isn't working is that
the game has either called Release() on it before drawing the crosshair,
or has otherwise cleared it.

Instead, create a new depth texture with MakeWTexture, and use it as an
additional render target to the directional lighting shader. In that
shader we simply copy the contents of _CameraDepthTexture onto it, then
we can inject it into the crosshair shader.

Only problem - it appears to be a mono texture as one of the eyes is
lined up, and the other isn't :-/