Montague's Mount: Force MakeWTexture to stereo
I don't think we can directly set it to stereo, so we rely on hacks.

Driver profile settings could probably do it, but we can also change it
to use the resolution of the depth target rather than the back buffer.

This has the side effect of allowing us to influence the creation mode
and force it to stereo that way.

The crosshair now correctly lines up in both eyes :)