Montague's Mount: NOTE STEREO CURSOR SHOW STOPPER :(
I was about to upload this when the directional shadows suddenly
disappeared and the stereo cursor went to the near clipping plane.

Somehow, injecting an extra render target seems to cause issues making
this approach a show stopper :(

Two places I've found consistently cause this:
1. Turn on the first generator, then walk back to the 2nd house and turn
on the light switch.

2. Load final save before end game. While walking towards stair case the
shadows disappear.

NFI why this is or what I can do about it.