Montague's Mount: Use DepthForce* to prevent the crosshair & shadows breaking
The issue with the stereo crosshair breaking turned out to be due to a
depth buffer being created at certain times with a resolution of
1024x1024, and since we are using DepthGetDepthResFrom = 2, that would
replace the helix mod depth buffer.

Since the depth buffer no longer matched the first render target, the
shader would not execute, shadows would disappear and the stereo
crosshair would stop working.

We can avoid this by using DepthGetDepthResFrom = 0, however that brings
us back to the mono depth buffer issue. Potentially this may be a way
forward if combined with the right StereoTextureEnable driver profile
setting to force it to stereo.

For now, use DepthForceWidth and DepthForceHeight to manually set the
correct depth buffer resolution. This will require the user to edit the
DX9Settings.ini file if they use a different resolution, but that seems
acceptable.