shadertool: Add auto fix for new UE3 lighting pattern seen in Life is Strange
Uses the same code as the existing UE3 shadow fix, substituting in the
right matrix.

Had to make the matrix register matching a little more relaxed to cope
with the swizzles used in this pattern.

Also redo the shader in Life is Strange with this script to make
grepping easier and prove that the script works.