mkrelease: Check for broken symlinks and abort if found
Since 3DMigoto is properly versioned and often updated I'm not including
a convenience copy in this repository like I did for Helix mod.

Rather, if you need to make a release of a game that uses it unpack the
appropriate 3Dmigoto release into a directory of the same name.

mkrelease.sh will now check and report any broken symlinks to prevent me
accidentally releasing a fix without bundling 3Dmigoto.