mkrelease: Remove bash dependency on pushd/popd by using subshell
pushd & popd are bash exclusives, preventing mkrelease from running on
other POSIX compliant shells. Instead, run the section of the script
that needs to change the working directory in a subshell, which prevents
the parent shell's working directory from changing and has the same
result as using pushd/popd, but now is standards compliant.