Introduce pyasm.py to implement shader assembly semantics in Python
This provides a Register class, which implements swizzles and masks
allowing the behaviour of these things to be replicated in Python.

Also implements a number of common shader instructions as python
functions that take Register instances as inputs and return new
Registers as outputs.

The idea is that some shader assembly code can be translated to and from
Python fairly easily with minimal syntax editing, allowing complicated
algorithms to be prototyped and debugged in Python before moving to real
assembly.