Witcher 3: Switch skipped shader to using a test in assembly
Use assembly here so that while the shader is enabled it will look like
the original. Dunno if assembly is strictly necessary here - I'm not
positive of what the issue with this shader was because it looks fine to
me, at least in the early part of the game.

Hook it up to the same toggle key as the other one, but rather than
setting the output to 0, use the discard instruction which has the same
visual result as handling=skip.