mkrelease: Allow to be used outside of repository & override date/version
Let's reuse this script for out of tree fixes as well. If --no-repo is
passed in instead of the tag name it will skip the dirty working tree
test and won't tag the commit.

Also allow a third parameter to be passed in to set a version number,
falling back to the date if nothing is passed in.