pyasm: Add py_to_asm function to convert a python routine into shader assembly
Write a function in Python with pyasm then use this to transform it into
assembly :)

This uses Python's built in code inspection and language parsing to
transform itself into assembly. This is probably overkill and will
likely cause me pain, but it was as much a learning exercise as anything
else.