Add feature to analyse how frames are constructed
This will dump out the context of every active 2D render target & depth
target after every draw call for the duration of one frame, allowing for
a shaderhacker to analyse in detail when and how an effect is drawn,
track down effects that are difficult to identify through traditional
hunting or are contracted through multiple stages.

This approach could very easily be extended to also dumping out all
texture inputs, which would also provide an ideal means to identify
texture hashes for filtering purposes.