Add scripts to copy driver profiles and clean up output
copy_profiles.sh copies a couple of files from the driver related to
profiles to a subdirectory. The idea is to grab these after a fresh
install for comparison between drivers if something breaks.

sanitise_nv_profiles.py takes the output of Geforce 3D Profile Manager
and sanitises the encoding so that any external tool can work with the
text. Currently this conversion is one way and cannot be be imported
back into the profile manager - this would not be difficult to add, but
is probably of much lower value than say, integrating profile management
into 3Dmigoto.