Add frame analysis options to select which output format to use
This adds a new option "analyse_options" which can enable dumping render
targets as jps, dds, one or the other, or both.

The default in the d3dx.ini is set to analyse_options = dump_rt_jps
since that requires the least disk space (by several orders of
magnitude), is the fastest and in many cases will be sufficient to track
down a shader that is simply drawing to the back buffer.

dump_rt can be used instead to try to dump out render targets as JPS
first, then fall back to DDS for render targets which cannot be saved as
JPS. This is useful for depth targets which are not saved as JPS, as
well as various other off-screen buffers such as shadow maps.

If someone is trying to analyse a particularly complicated render path
they can choose to enable dump_rt_dds instead (or in conjunction with
dump_rt_jps) to make sure they get DDS files of all render targets.