Witcher 3: Fix MSAA hairworks with fake_o0
This removes the workaround from 0675778425b068a0 (which copies the
resolved hairworks depth buffer to the regular depth buffer), and fixes
the one-eye rendering issue in 4f47213227634ea5 (which resolves the MSAA
hairworks depth buffer into a non-MSAA depth buffer).

The shader responsible for resolving the hairworks MSAA depth buffer
was only being run for the right eye, creating a one-eye rendering
issue. By assigning a dummy render target when this shader is run the
driver's heuristics will run it twice and the one-eye issue is no more.

A similar approach may well apply to other games with one-eye issues -
the frame_analysis feature was crucial in tracking down the correct
shader and may help elsewhere.