Change default hunting toggle key to numpad 0 instead of insert
Insert conflicts with the Ctrl+Alt+Insert and could lead regular users
to turn on hunting mode when they try to disable the driver overlay,
leading to confusion. Change the default to numpad 0, which is far less
likely to be used by chance and still convenient as it is next to the
rest of the hunting buttons.