Make hunting=0 disabled again, use hunting=2 for soft disabled
The idea is to distinguish between a user playing a game who won't ever
want hunting to be enabled, and a shaderhacker who might want to play
with it disabled, but turn it on on the fly.

hunting=0 = hard disabled. For now this still has enough enabled that
	    turning hunting on in the d3dx.ini file and reloading it
	    will function the same as toggle_hunting does today.

hunting=1 = enabled

hunting=2 = soft disabled. The toggle_hunting key can be used to turn it
            on at any time.

- Also enable the live d3dx.ini reload support by default. This is now
  pretty well tested and no longer experimental, and gives a
  shaderhacker the ability to enable hunting without restarting the game
  even if they did launch with hunting=0.

- Enables show_original on F9 by default. This is also now a well
  supported and very useful feature. Previously the comment had this set
  to F11, but F9 is easier to find by touch on many keyboards that have
  a gap between F8 and F9, which is useful to avoid taking your eyes off
  the screen while watching for changes.

- Fix a bug which would have prevented index buffer filtering from
  working with hunting disabled (untested).