Lichdom: Use row_major keyword to stop fxc transposing matrices
This seems backwards to me - although the matrices are declared
row_major, the fact that they use dot products during the
multiplications suggests that they are in fact column major.

Did something change for DX11 (even if just the terminology) or have I
got mixed up somewhere along the lines?