Lichdom: Fix snow noise & occlusion
Using frame analysis I was able to find that the view projection matrix
was inside one of the (unused) constant buffers that was still bound to
the pipeline.

Similar to the decals, inverse and copy that matrix in to the pixel
shader where it is used to correct a world space coordinate.