Lichdom: Approximate fix for blue light from distance
This was tricky - the VP matrix from the vertex shader turned out to be
the wrong one for the correction, but the correct one is still bound so
add the definition and use it.

Up close I was able to make it line up perfectly, but from a distance
it's slightly out. Need to check what happens elsewhere in the game -
this might be rotation dependent.