Lichdom: Accurate fix for CryEngine directional shadows!!!
Achievement unlocked: World first fix for CryEngine shadows ;-)

These were tricky to solve, but at the end of the day the pattern is
actually not that hard.

The far clipping plane and screen resolution needs to be copied from the
vertex shader, as the variable in the PS that suggests it might have the
far clipping plane is not the right one - seems to be for the shadow
map?

The SV_Position v0 is adjusted in the pixel shader based on the
information from the depth buffer. This is scaled back to world Z by
multiplying by the far clipping plane, then the stereo correction
formula (with perspective divide) is applied and finally multiplied by
the screen resolution.