Lichdom: Fix more light/shadow shaders
Fixes the light around shield recharge orbs

Also fixes another pixel shader that is paired with the same VS, but not
sure what light it is. There's a few more likely to need a similar fix,
but the pattern isn't quite the same so I'll wait until I actually find
one.