Lichdom: Disable distortion effects when W > 5 to avoid pulsing
The distortion effects in this game pulsate. It seems to be a problem
with the refraction sampler, but so far I'm unable to track it down. For
now, kill the distortion effects from a distance, so we still get the
distortion up close but avoid seeing the pulsating effect from further
away.