Lichdom: First pass adjusting damage indicators
Adjusts text (matched on VS+PS pair) only above a certain point on the
screen since we don't want to adjust the UI that sits in front of the
hands.

Disabled by default since it messes up the menus - press X to enable.

Need to see if I can implement texture filtering to improve this.