Lichdom: Kill pulsating transparency on Roth / Dragon vision
Looks like the environment refraction texture does not cover enough of
the screen to be useful in 3D, leading to a pulsating effect near the
edges.

For now, kill the transparency. Doesn't look great, but better than
pulsating. I need to track down where the sampler comes from see if I
can increase it's coverage.