Lichdom: Fix transparent refraction effects with expand_region_copy
These effects were broken due to the game calling CopySubresourceRegion
with a 2D rectangle around the effect that did not cover the entire
effect in 3D.

I've added a new feature to 3DMigoto to expand the region in these
copies to ensure that the entire effect is covered, which should fix
this.

Leaving the F3 toggle in for now (inverted the starting state) as I'm
not convinced the texture hashes are reliable - while testing at one
point I ended up with the texture being 1904x1002 instead of 1920x1080
(possibly by switching to windowed mode and back?), and I need to
confirm that the matches back buffer logic works these days.