Lichdom: Fix issue with objects being falsely culled while visible
This depends on a new feature scheduled for 3DMigoto 1.1.32, which can
deny the game from reading it's own depth buffer.

CryEngine reads information (depth buffer?) back from the GPU to
determine which objects were drawn, but since that information has been
adjusted for stereo the game will falsely cull objects that it thinks
were not drawn but actually were.

Using the new deny_cpu_read feature, 3DMigoto gives the game a blank
buffer instead of the one from the GPU, which prevents it from being
able to make culling decisions and everything is drawn :)