Lichdom: Fix lights clipping at bounds of shadow map
Found a cave in the snow level (the one with the loot machine, 2nd
stage) where the directional(?) shadows cut off in a line. It's bogus,
but when standing in line with the edge of the shadow it would clip
badly, which looks worse.

Adjust the bounding box of shadow inside the vertex shader to prevent
this clipping. Compensate for the adjustment by also adjusting the
texcoord and world position (which needed to be multiplied by the far
clipping plane to be accurate).