Lichdom: Use partner shader filtering to avoid breaking decals when fixing wet rocks
The previous fix broke decals. Filter the adjustment based on the active
pixel shader so that it is only done for wet rocks, and not for decals.

This is partner shader filtering passing the value into the shader, but
it is not officially supported in 3DMigoto yet - I'm abusing the fact
that the vertex shader overrides are processed before the pixel shader
overrides on each draw call to achieve my goals. Once officially
supported, the duration of y2 in the pixel shader will be explicitly set
to a single draw call.

Possibly could use depth buffer filtering instead, but I haven't
extended that to be passed into the shader yet, and haven't checked if
the depth buffer is active for either/both of these.