Far Cry 4: Fix patchy snow inverted emboss
Spotted something wrong with the ground near the final fortress while
using show_original to check soft shadows - seems the snow patches were
inverted so they sank into the ground instead of sitting on top of it.

Latest decompiler improvements didn't help. Would have redone it in
assembly, but not without finding this shader used in a reflection to
verify the fix still works.

Solution was to remove the negate. I believe this is due to differences
in the result of boolean operations in assembly vs HLSL:

In assembly:
   r0.w = 0 < r0.z;
means r0.w will contain either 0x00000000 or 0xffffffff.
-r0.w in the itof instruction will treat r0.w as a signed integer and
take the 2's complement:
  0x00000000 -> 0xffffffff + 1 = 0x00000000
  0xffffffff -> 0x00000000 + 1 = 0x00000001
then convert that to a float:
  0x00000000 -> 0x00000000
  0x00000001 -> 0x3f800000 (1.0)

But, in HLSL a boolean operation works differently:
   r0.w = 0 < r0.z;
will result in r0.w containing either 0 or 1, which differs from
assembly, and from there on everything is backwards, the 2's complement:
  0x00000000 -> 0xffffffff + 1 = 0x00000000
  0x00000001 -> 0xfffffffe + 1 = 0xffffffff
then convert that to a float:
  0x00000000 -> 0x00000000
  0xffffffff -> 0xbf800000 (-1.0)

A more general solution might be to change all the boolean operations to
return 0xffffffff or -1. Needs testing to confirm that wouldn't break
anything else though.