Crysis 3: Move environment reflections to infinity (high quality)
Use a different adjustment amount for the environment reflection to move
it to infinity instead of water depth. Much nicer to look at now.

Not sure if this is 100% accurate - having a hard time telling if it's
at infinity or near it, but if it isn't accurate it's close enough.

Diff looks more complex than it is as I synchronised the
ripple/non-ripple shaders with each other and renamed a variable.