Mad Max: Fix projected lights in second stronghold
This is a hybrid of two patterns we have already seen - needs the
inverse view-projection matrix to fix v2, as well as the modified stereo
correction on r0 (from v0) and a halo fix in the vertex shader.

The halo fix broke some other lights sharing the same vertex shader,
which also needed the same pattern in the pixel shader to fix again.