Mad Max: Adjust bloom from depth buffer (experimental)
This uses the variation of my auto UI depth adjustment I was playing with
for Witcher 3. The depth buffer is injected into the bloom/lens flare
vertex shader and used to determine what depth to draw the bloom at.

Each corner of the bloom is calculated independently (and each eye is
calculated independently), which can lead to a bloom blob stretching
depending on what surrounds the light source on the depth buffer, but it
isn't terrible.