Mad Max: Fix bloom/lens flare opacity being calculated from wrong spot on depth buffer
Calculating the opacity from the wrong spot on the depth buffer leads to
the bloom & lens flare appearing & disappearing when they should not,
being too pronounced, or not visible at all when it should be.

The pixel shader has the depth of the light source in Z buffer scale,
which we can use to correct this. Scale it out to world Z and use it to
offset the location of the depth buffer read to determine the opacity of
the bloom & lens flare.

Debugging is included to show where the depth buffer is being sampled.