Witcher 3: Remove conflicting convergence presets
From Helifax:
> EDIT: Forgot to mention I added in the ini file some presets on F2, F3
> (For normal and cinematics modes). I think those can be removed;)) or
> if you like them they can go inside;)) Attachments

We're already using F3 to disable water reflections on SLI systems
(which might be worth checking if an update has fixed), so I opted to
remove these for now. We can probably add a preset back - will need to
give it a try in the game to see what works best.