Witcher 3 experimental auto UI depth: Use W-buffer for depth
The very first thing this game does in it's post-processing stage is
render a W-buffer, which is similar to a depth buffer, but pre-scaled to
world coordinates.

Use the new resource copying support to copy into all the HUD shaders
instead of the depth buffer, then replace the scaling code to just
return the W value from the buffer directly.