Witcher 3: Fix misaligned UI elements (mouse, some icons)
Some of the UI elements in The Witcher 3 are drawn to a small offscreen
buffer before being copied to the screen. The depth adjustment was being
applied in both of these stages, which caused some UI elements such as
the mouse cursor and certain icons to become slightly misaligned with
the rest of the UI.

Use render target size filtering to only adjust UI elements when they
are drawn to the screen to fix this issue.