Viscera Cleanup Detail: Adjust bloom from depth buffer (UE3 auto stereo code!)
Since some bloom in UE3 does not have access to the depth information
(neither the vertex or pixel shader), adapt my auto stereo crosshair
code to work in UE3, and update it to work with arbitrary coordinates
like the DX11 version.

Take the depth buffer from all the shadow shaders. The bloom pixel
shader itself has access to this, but I noticed that it is blank on the
bloom that already renders at the correct depth, and I didn't want to
risk ending up with no depth information.

The coordinates of the center of the effect do not appear to be
available unfortunately, so adjust each vertex individually using their
own coordinates.

Ignore depth buffer reads returning 1.0, which are bogus (e.g. due to
clipping at the edge of the screen leaving the depth buffer blank).